# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "spec_text": {
            "maxLength": 50000,
            "minLength": 10,
            "title": "Spec Text",
            "type": "string"
          },
          "submittal_text": {
            "maxLength": 50000,
            "minLength": 10,
            "title": "Submittal Text",
            "type": "string"
          },
          "system_id": {
            "default": "CUSTOM",
            "maxLength": 50,
            "title": "System Id",
            "type": "string"
          }
        },
        "required": [
          "spec_text",
          "submittal_text"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "CopilotQuery": {
        "properties": {
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "CopilotQuery",
        "type": "object"
      },
      "CreateCaseRequest": {
        "properties": {
          "name": {
            "default": "",
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "CreateCaseRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebhookSubscribeRequest": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "WebhookSubscribeRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Spec-to-Site Deviation Sentinel for hyperscale data-centre EPC delivery",
    "title": "Pramaan — EPC Deviation Intelligence",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyze": {
      "post": {
        "operationId": "v1_analyze",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Analyze",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/cases": {
      "post": {
        "description": "Create a new case. The returned `secret` is shown exactly once — it\nis the only credential for every /cases/{case_id}/* route on this case\nand cannot be recovered if lost, the same way a demo token can't be.",
        "operationId": "v1_create_case",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Create Case",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/cases/{case_id}/webhooks": {
      "delete": {
        "operationId": "v1_clear_case_webhooks",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Clear Case Webhooks",
        "tags": [
          "v1"
        ]
      },
      "get": {
        "operationId": "v1_get_case_webhooks",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Case Webhooks",
        "tags": [
          "v1"
        ]
      },
      "post": {
        "operationId": "v1_subscribe_case_webhook",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookSubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Subscribe Case Webhook",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/copilot": {
      "post": {
        "operationId": "v1_copilot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Copilot",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/jobs/analyze": {
      "post": {
        "description": "Submit an analysis as a background job; returns immediately (202) with a\njob_id to poll. Same auth + analysis rate limit as /analyze.",
        "operationId": "v1_submit_analyze_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Submit Analyze Job",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/webhooks": {
      "delete": {
        "description": "Retired global mutation endpoint retained as an explicit 410.",
        "operationId": "v1_clear_webhooks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Clear Webhooks",
        "tags": [
          "v1"
        ]
      },
      "get": {
        "operationId": "v1_get_webhooks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Webhooks",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/webhooks/subscribe": {
      "post": {
        "operationId": "v1_subscribe_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookSubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Subscribe Webhook",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}